Version bumps and build-tests for AUR.
Note: this package is slightly different from the others.
It doesn't pull in any code modified by the elementary team.
It pulls in the bzr tip of libappindicator (from Canonical)
and patches the .desktop for pantheon sessions.

This is an alternative approach that might make some indicators
work in Arch when their pantheon-patched versions do not.