Setting backend to gsettings in user configuration.
This is technically a workaround: the real problem is that
compiz should be functional no matter what backend is in use, yet
it refuses to load or allow the rest of the desktop to load unless
it's using gsettings.

If you were previously using compiz this is going to break your
configuration. You can re-import your old config, or use mine.
compiz-qq.profile covers the custom settings I use for pantheon3d

It is not reccomended that you run a compiz desktop without a GPU!